+++ node-video-crime-miner/test/unit_tests/Rekognition/videoLabelUtils.test.ts
